+++ modules/utils/elements.js
@@ -1,0 +1,2 @@
+export const addBtn = document.querySelector('.btn-add');
+export const bookSection = document.querySelector('.display-books');
