+++ modules/utils/navControl.js
@@ -1,0 +1,7 @@
+const navControl = (show, firstHide, secondHide) => {
+  show.classList.remove('d-none');
+  firstHide.classList.add('d-none');
+  secondHide.classList.add('d-none');
+};
+
+export default navControl;
